move interpolation from propspeed
