+++ dipy/core/interpolation.pyx
@@ -1,5 +1,6 @@
 cimport cython
 cimport numpy as np
+cimport numpy as cnp
 
 import numpy as np
 import time
@@ -7,6 +8,35 @@ import time
 from libc.math cimport floor
 
 from dipy.align.fused_types cimport floating, number
+
+
+@cython.cdivision(True)
+cdef cnp.npy_intp offset(cnp.npy_intp *indices,
+                         cnp.npy_intp *strides,
+                         int lenind,
+                         int typesize) nogil:
+    """ Access any element of any ndimensional numpy array using cython.
+
+    Parameters
+    ------------
+    indices : cnp.npy_intp * (int64 *)
+        Indices of the array for which we want to find the offset.
+    strides : cnp.npy_intp * strides
+    lenind : int, len(indices)
+    typesize : int
+        Number of bytes for data type e.g. if 8 for double, 4 for int32
+
+    Returns
+    ----------
+    offset : integer
+        Element position in array
+    """
+    cdef int i
+    cdef cnp.npy_intp summ = 0
+    for i from 0 <= i < lenind:
+        summ += strides[i] * indices[i]
+    summ /= <cnp.npy_intp>typesize
+    return summ
 
 
 cdef void splitoffset(float *offset, size_t *index, size_t shape) nogil:
@@ -71,6 +101,118 @@ def trilinear_interp(np.ndarray[np.float32_t, ndim=4, mode='strided'] data,
                 for LL from 0 <= LL < last_d:
                     result[LL] += data[x_ind+ii,y_ind+jj,z_ind+kk,LL]*weight
     return result
+
+
+@cython.boundscheck(False)
+@cython.wraparound(False)
+def map_coordinates_trilinear_iso(cnp.ndarray[double, ndim=3] data,
+                                  cnp.ndarray[double, ndim=2] points,
+                                  cnp.ndarray[cnp.npy_intp, ndim=1] data_strides,
+                                  cnp.npy_intp len_points,
+                                  cnp.ndarray[double, ndim=1] result):
+    """ Trilinear interpolation (isotropic voxel size)
+
+    Has similar behavior to ``map_coordinates`` from ``scipy.ndimage``
+
+    Parameters
+    ----------
+    data : array, float64 shape (X, Y, Z)
+    points : array, float64 shape(N, 3)
+    data_strides : array npy_intp shape (3,)
+        Strides sequence for `data` array
+    len_points : cnp.npy_intp
+        Number of points to interpolate
+    result : array, float64 shape(N)
+        The output array. This array should be initialized before you call
+        this function.  On exit it will contain the interpolated values from
+        `data` at points given by `points`.
+
+    Returns
+    -------
+    None
+
+    Notes
+    -----
+    The output array `result` is filled in-place.
+    """
+    cdef:
+        double w[8]
+        double values[24]
+        cnp.npy_intp index[24]
+        cnp.npy_intp off, i, j
+        double *ds=<double *> cnp.PyArray_DATA(data)
+        double *ps=<double *> cnp.PyArray_DATA(points)
+        cnp.npy_intp *strides = <cnp.npy_intp *> cnp.PyArray_DATA(data_strides)
+        double *rs=<double *> cnp.PyArray_DATA(result)
+
+    if not cnp.PyArray_CHKFLAGS(data, cnp.NPY_C_CONTIGUOUS):
+        raise ValueError(u"data is not C contiguous")
+    if not cnp.PyArray_CHKFLAGS(points, cnp.NPY_C_CONTIGUOUS):
+        raise ValueError(u"points is not C contiguous")
+    if not cnp.PyArray_CHKFLAGS(data_strides, cnp.NPY_C_CONTIGUOUS):
+        raise ValueError(u"data_strides is not C contiguous")
+    if not cnp.PyArray_CHKFLAGS(result, cnp.NPY_C_CONTIGUOUS):
+        raise ValueError(u"result is not C contiguous")
+    with nogil:
+        for i in range(len_points):
+            _trilinear_interpolation_iso(&ps[i * 3],
+                                         <double *> w,
+                                         <cnp.npy_intp *> index)
+            rs[i] = 0
+            for j in range(8):
+                weight = w[j]
+                off = offset(&index[j * 3], <cnp.npy_intp *> strides, 3, 8)
+                value = ds[off]
+                rs[i] += weight * value
+    return
+
+
+cdef void _trilinear_interpolation_iso(double *X,
+                                       double *W,
+                                       cnp.npy_intp *IN) nogil:
+    """ Interpolate in 3d volumes given point X
+
+    Returns
+    -------
+    W : weights
+    IN : indices of the volume
+    """
+
+    cdef double Xf[3]
+    cdef double d[3]
+    cdef double nd[3]
+    cdef cnp.npy_intp i
+    # define the rectangular box where every corner is a neighboring voxel
+    # (assuming center) !!! this needs to change for the affine case
+    for i from 0 <= i < 3:
+        Xf[i] = floor(X[i])
+        d[i] = X[i] - Xf[i]
+        nd[i] = 1 - d[i]
+    # weights
+    # the weights are actually the volumes of the 8 smaller boxes that define
+    # the initial rectangular box for more on trilinear have a look here
+    # http://en.wikipedia.org/wiki/Trilinear_interpolation
+    # http://local.wasp.uwa.edu.au/~pbourke/miscellaneous/interpolation/index.html
+    W[0]=nd[0] * nd[1] * nd[2]
+    W[1]= d[0] * nd[1] * nd[2]
+    W[2]=nd[0] *  d[1] * nd[2]
+    W[3]=nd[0] * nd[1] *  d[2]
+    W[4]= d[0] *  d[1] * nd[2]
+    W[5]=nd[0] *  d[1] *  d[2]
+    W[6]= d[0] * nd[1] *  d[2]
+    W[7]= d[0] *  d[1] *  d[2]
+    # indices
+    # the indices give you the indices of the neighboring voxels (the corners
+    # of the box) e.g. the qa coordinates
+    IN[0] =<cnp.npy_intp>Xf[0];   IN[1] =<cnp.npy_intp>Xf[1];    IN[2] =<cnp.npy_intp>Xf[2]
+    IN[3] =<cnp.npy_intp>Xf[0]+1; IN[4] =<cnp.npy_intp>Xf[1];    IN[5] =<cnp.npy_intp>Xf[2]
+    IN[6] =<cnp.npy_intp>Xf[0];   IN[7] =<cnp.npy_intp>Xf[1]+1;  IN[8] =<cnp.npy_intp>Xf[2]
+    IN[9] =<cnp.npy_intp>Xf[0];   IN[10]=<cnp.npy_intp>Xf[1];    IN[11]=<cnp.npy_intp>Xf[2]+1
+    IN[12]=<cnp.npy_intp>Xf[0]+1; IN[13]=<cnp.npy_intp>Xf[1]+1;  IN[14]=<cnp.npy_intp>Xf[2]
+    IN[15]=<cnp.npy_intp>Xf[0];   IN[16]=<cnp.npy_intp>Xf[1]+1;  IN[17]=<cnp.npy_intp>Xf[2]+1
+    IN[18]=<cnp.npy_intp>Xf[0]+1; IN[19]=<cnp.npy_intp>Xf[1];    IN[20]=<cnp.npy_intp>Xf[2]+1
+    IN[21]=<cnp.npy_intp>Xf[0]+1; IN[22]=<cnp.npy_intp>Xf[1]+1;  IN[23]=<cnp.npy_intp>Xf[2]+1
+    return
 
 
 @cython.boundscheck(False)
